SoilifyAPI/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
bcrypt==4.1.1

# Google Gemini AI (NO LangChain)
google-generativeai==0.3.1

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# CORS
starlette==0.27.0

# Email (for password reset if needed)
python-jose==3.3.0

# Date/Time utilities
python-dateutil==2.8.2

# Logging
colorlog==6.8.0